Fixed duplicate TimeStamps when deleting
For a reason unbeknown to me, sometimes when an item was removed via the
function `notifyItemRemoved` within the Adapter, not all of the items in
the RecyclerView were being updated, thus producing the appearance of
duplicate items. That is strange behavior because that same function
operates as expected in other apps. The use of `notifyItemRangeRemoved`
resolves the error.

Side note, this error is not due to the version of RecyclerView in use
is a beta version. I attempted to fix the error by reverting back to a
stable version of RecyclerView, regardless, the error remained.