Removed a hack that tried to fix a now resolved issue
I refactored to `notifyItemRangeRemoved` function, from `notifyItemRemoved`, to fix an issue with duplicate list items. That issue has been resolved. I need to revert back to `notifyItemRemoved`, because `notifyItemRangeRemoved` is causing another issue. When a TimeStamp is deleted, RecyclerView automatically jumps to the top of the list because it thinks the entire list was updated.